requirements/test.txt
-r base.txt

factory_boy
flake8
pytest
pytest-django
pytest-watch
